#321 Added subdomain column to EnsemblAssembly.
