+++ rnacentral/portal/management/commands/update_ensembl_assembly.py
@@ -12,9 +12,63 @@ limitations under the License.
 """
 from __future__ import print_function
 
+from collections import defaultdict
+
+import pymysql.cursors
 from django.core.management.base import BaseCommand
 from portal.models import EnsemblAssembly
-from .update_ensembl_genome_mapping import get_ensembl_connection, get_ensembl_databases
+
+
+def get_ensembl_connection():
+    """Connect to the public Ensembl MySQL database."""
+    return pymysql.connect(
+        host='ensembldb.ensembl.org',
+        user='anonymous',
+        port=3306,
+        cursorclass=pymysql.cursors.DictCursor
+    )
+
+
+def get_ensembl_genomes_connection():
+    """
+    Connect to the public Ensembl Genomes MySQL database. See:
+    http://ensemblgenomes.org/info/access/mysql
+    """
+    return pymysql.connect(
+        host='mysql-eg-publicsql.ebi.ac.uk',
+        user='anonymous',
+        port=4157,
+        cursorclass=pymysql.cursors.DictCursor
+    )
+
+
+def get_ensembl_databases(cursor):
+    """
+    Get a list of all available databases.
+    Return a list of the most recent core databases, for example:
+        homo_sapiens_core_91_38
+        mus_musculus_core_91_38
+    """
+    databases = defaultdict(list)
+    cursor.execute("show databases")
+    for result in cursor.fetchall():
+        database = result['Database']
+        if database.count('_') != 4 or 'mirror' in database:
+            continue
+        genus, species, database_type, ensembl_release, _ = database.split('_')
+        ensembl_release = int(ensembl_release)
+        if ensembl_release < 80:
+            continue
+        if database_type == 'core':
+            organism = genus + ' ' + species
+            databases[organism].append((ensembl_release, database))
+
+    to_analyse = []
+    # get the most recent database
+    for organism, dbs in databases.iteritems():
+        most_recent = max(dbs, key=lambda item: item[0])
+        to_analyse.append(most_recent[1])
+    return to_analyse
 
 
 example_locations = {
@@ -121,6 +175,24 @@ example_locations = {
 }
 
 
+def domain_url(division):
+    """Given E! division, returns E!/E! Genomes url."""
+    if division == 'Ensembl':
+        subdomain = 'ensembl.org'
+    elif division == 'Ensembl Plants':
+        subdomain = 'plants.ensembl.org'
+    elif division == 'Ensembl Metazoa':
+        subdomain = 'metazoa.ensembl.org'
+    elif division == 'Ensembl Bacteria':
+        subdomain = 'bacteria.ensembl.org'
+    elif division == 'Ensembl Fungi':
+        subdomain = 'fungi.ensembl.org'
+    elif division == 'Ensembl Protists':
+        subdomain = 'protists.ensembl.org'
+
+    return subdomain
+
+
 def get_ensembl_metadata(cursor, database):
     """Get metadata about genomic assemblies used in Ensembl."""
     cursor.execute("USE %s" % database)
@@ -169,6 +241,7 @@ def store_ensembl_metadata(metadata):
         taxid=metadata['species.taxonomy_id'],
         ensembl_url=metadata['species.url'].lower(),
         division=metadata['species.division'],
+        subdomain=domain_url(metadata['species.division']),
         example_chromosome=example_location['chromosome'],
         example_start=example_location['start'],
         example_end=example_location['end']
@@ -184,6 +257,7 @@ def store_ensembl_metadata(metadata):
         taxid=metadata['species.taxonomy_id'],
         ensembl_url=metadata['species.url'].lower(),
         division=metadata['species.division'],
+        subdomain=domain_url(metadata['species.division']),
         example_chromosome=example_location['chromosome'],
         example_start=example_location['start'],
         example_end=example_location['end']
